Fixes because typo and rating @<player> syntax
